Fixing project structure
