+++ lib/app/modules/About/bindings/about_binding.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:ultimate_alarm_clock/app/modules/settings/controllers/about_controller.dart';
+import 'package:ultimate_alarm_clock/app/modules/About/controller/about_controller.dart';
 import 'package:ultimate_alarm_clock/app/modules/settings/controllers/theme_controller.dart';
 
 class AboutBinding extends Bindings {
